apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
prometheus-fastapi-instrumentator==7.0.0
websockets==14.1
httpx==0.28.1
pandas==2.2.3
numpy==2.1.3
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1
loguru==0.7.3
ccxt==4.4.36
# pandas-ta==0.3.14b0  # Temporarily disabled
scikit-learn==1.6.0
lightgbm==4.5.0
aiofiles==24.1.0
stripe==11.2.0
aiohttp==3.10.11
requests==2.32.3

